requirements.txt
streamlit
sqlite3
Pillow
